requirements.txt
alembic==1.4.1
appdirs==1.4.3
atomicwrites==1.4.0
attrs==19.3.0
awscli==1.18.55
Babel==2.8.0
botocore==1.16.5
cffi==1.17.1
Click==7.0
colorama==0.4.3
cryptography==44.0.2
distlib==0.3.0
docutils==0.15.2
ecdsa==0.15
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jwt==1.3.1
Mako==1.1.2
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.9.10
py==1.8.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.1.1
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==1.1.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.3
PyYAML==5.3.1
rsa==3.4.2
s3transfer==0.3.3
setuptools==78.1.0
six==1.17.0
SQLAlchemy==1.3.13
urllib3==1.25.9
virtualenv==20.0.10
wcwidth==0.1.9
Werkzeug==1.0.0
WTForms==2.2.1
